Use a nested path to prevent hitting the directory limitation
EXT3 has a limit on the number of files that a directory can hold. To
prevent us hitting this limit we'll distribute our files to a depth of
two directories.

e.g. If our BSON ID is '5124bd9a686c8228a9000005', then the asset path
generated would be: /51/24/5124bd9a686c8228a9000005/