Replace state_machine gem with state_machines-mongoid
The old gem seems to be unsupported and raises errors about protected
`around_validation` methods being called.  The new gem is a fork of the
original with all the integrations separated out.  Seems a safer option
than changing the gem to one of the other state machine gems as at least
this way we don't have to change the API.