Use strong parameters not attr_accessible
User had attr_accessible, but was not actually a target for any writes.  Asset
didn't have attr_accessible and was a target for writes, but it's not clear
which attributes should be available to those endpoints.  Going by the
documentation in gds-api-adapters it's just file, even though access_limited
and organisation_slug appear to be available.  Interestingly the specs tried
to set title atttributes, but that's not a declared field so mongoid rejects
the update.