Use environment variables for secret MongoDB config
~~12 Factor style~~

This also moves the `retry_interval` out of the `alphagov-deployment`
version of `mongoid.yml`, as it's not a secret, and now we're using
ENV vars for connection details, we don't need a separate/secret
production version of the config.

Convert `development`/`test` to use `uri` based config for consistency
and drop the port, as this doesn't seem to be need - i assume mongo
defaults to that port.