Add hook to save assets to cloud storage as well as NFS
The `S3Storage#save` method is called when an asset is virus scanned
successfully. At the moment the `S3Storage#save` method is a no-op but
this commit ensures that the method is called when an asset is marked as
clean.

We're planning to treat files that are uploaded to S3 as disposable
while we work to get this functionality integrated. For that reason,
and to have a minimal impact on the application, we are not storing
the "uploaded" state in the database.