Call S3Uploader#upload when an uploaded file is clean
At the moment the `upload` method is a no-op but this commit ensures
that the method is called when an asset is marked as clean.

We're planning to treat files that are uploaded to S3 as disposable
while we work to get this functionality integrated. For that reason,
and to have a minimal impact on the application, we are not storing
the "uploaded" state in the database.