Upload files to s3
This commit adds an implementation of S3Uploade#upload that uploads
the asset's file to the S3 bucket specified in the BUCKET_NAME
environment variable.

The default names for the other environment variables that `aws-sdk`
relies on (AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, AWS_REGION) are
assumed to be set at run time.

Now that S3Uploader#upload has a side-effect we also add a global stub
for the method to ensure that it is not called when other specs are
run.