Upgrade rspec-rails from v3.3.3 -> v3.4.2
I changed the constraint on `rspec-rails` from `~> 3.3.0` to `~> 3.4.0`
and ran `bundle update --conservative rspec-rails`. This upgraded
`rspec-core` from v3.3.2 to v3.4.4 which fixes the following deprecation
warning which was appearing multiple times when I ran the specs using
`bundle exec rake` or when the `asset-manager-worker` process was
started on my developer VM.

    [DEPRECATION] `last_comment` is deprecated.
    Please use `last_description` instead.

This issue [1] suggested that the deprecation warnings could be fixed
by upgrading rspec-core to at least v3.4.4 which this commit does.

c.f. alphagov/manuals-publisher#827

[1]: ruby/rake#133