Notify Errbit via Airbrake if upload to S3 fails
The uploads to S3 happen in a `Delayed::Job`. While we are working
on configuring S3 in the various deployment environments, if this job
fails it may be due to a configuration error. In that case we'd like
to know what the exception was so that we can fix it. This commit
notifies Errbit (via the Airbrake library) and re-raises the exception
so that `Delayed::Job` will retry it (in case we fix the problem, or it
was temporary).

Note that this change replicates the pattern used in
`Asset#scan_for_viruses`, the other method invoked via `Delayed::Job`.
A better solution would be to implement a generic solution for reporting
all failures that occur in a `Delayed::Job`. See #68 for suggestions on
how to implement this.