Notify Airbrake / Errbit if S3 Upload fails
The uploads to S3 happen in a `delayed_job` job. While we are working
on configuring S3 in the various deployment environments, if this job
fails it may be due to a configuration error. In that case we'd like
to know what the exception was so that we can fix it. This commit
notifies Errbit (via the Airbrake library) and re-raises the exception
so that delayed job will retry it (in case we fix the problem, or it
was temporary).