Optionally stream asset from S3 on public URL
This is a very naïve approach to serving assets stored on S3 (instead of
on NFS) from the Asset Manager app. The current behaviour should be
unchanged and so this should be safe to deploy to production.

The new behaviour is triggered by adding "stream_from_s3=true" to the
URL query string.

The default behaviour uses `ActionController::DataStreaming#send_file`
[1] in combination with the `config.action_dispatch.x_sendfile_header`
option and the nginx configuration so that the asset is actually
streamed directly from disk by nginx and does not pass through the Rails
app at all.

The new behaviour downloads the file from S3 into a StringIO object in
memory and then uses `ActionController::DataStreaming#send_data` [2] to
send the data to the browser.

The new approach is unlikely to be viable in production, because it
means that all data for all assets has to be downloaded from S3,
temporarily stored in memory, and then streamed to the browser from the
Rails app, tying up more memory and a Ruby thread/process each time.
This could easily mean that the app is overloaded and requests are
blocked or result in a time-out.

However, we think the new behaviour is a useful first step to verify
that the app is correctly wired up to use the relevant AWS credentials,
etc.

The only way the new behaviour differs from the default behaviour from
the client browser's point of view is that an `ETag` header is sent in
the response. This is because `Rack::ETag` [3] in the Rails middleware
stack automatically sets an `ETag` header on all String bodies. Since
in the new approach the file response is sent as a String from the Rails
app passing through the middleware stack and it does not respond to the
`to_path` method, an `ETag` header is added.

It's not obvious whether this is actually a problem, but we should
definitely investigate further if we want to pursue using the new
behaviour as the default behaviour.

[1]: http://api.rubyonrails.org/v4.2.7.1/classes/ActionController/DataStreaming.html#method-i-send_file
[2]: http://api.rubyonrails.org/v4.2.7.1/classes/ActionController/DataStreaming.html#method-i-send_data
[3]: http://www.rubydoc.info/gems/rack/1.6.4/Rack/ETag