Avoid connecting to AWS if S3 bucket name not set
This commit introduces `S3Storage.build` to as a way for
`Services.cloud_storage` to construct the service object. If the bucket
name is not set then an instance of the new `S3Storage::Null` class is
returned. Calls to methods on this instance do not attempt to connect to
AWS and hence we should avoid lots of exceptions being reported to
Errbit.

We're making an assumption here that if the bucket name is set then it's
very likely that the other AWS environment variables will be set.