Upgrade Mongoid 4.0.2 -> 5.2.1
Upgrade Mongoid to the latest patch release in the 5 series using
`bundle update --conservative mongoid`.

I used `rails g mongoid:config` to update `config/mongoid.yml` and
modified the configuration to use the correct database connection
URI. I created a production config based on other mongoid-based
applications in the organisation, in particular following the pattern
we used for upgrading Mongoid from 4 to 5 in Manuals Publisher[1]

In Mongoid 5 documents are no longer ordered by `_id` by default[2]. I've
changed the specs that previously relied on the implicit ordering by
`_id` when calling the `last` method to make this ordering
explicit. There's probably a more intention-revealing way to write
these expectations, but doing it this way for now makes it clear why
the change has to be made - we can improve it later.

[1] https://github.com/alphagov/manuals-publisher/pull/1127
[2] https://docs.mongodb.com/mongoid/master/tutorials/mongoid-upgrade/