Extract AWS_S3_BUCKET_NAME env var into app config
So it's easier to stub.

Unfortunately I've had to disable a Rubocop rule around one line. This
violation is due to the digit `3` in the variable name. It looks as if
the constraints for this rule have been relaxed [1] in Rubocop v0.44.0,
but we're currently using v0.43.0 and upgrading means first upgrading
the constraints on the Rubocop version in the specification for the
`govuk-lint` gem and then upgrading that gem in this project. I'm
going to do that separately and I'll make a note to remove this
disabling of the rule when I do that.

[1]: bbatsov/rubocop@833aad6ccb35a16c6196830e936f4ef34384590b