Error if asset requested from S3, but S3 not configured
If the AWS S3 bucket is not configured and a request is made to stream
an asset from S3, then the server now responds with a
"500 Internal server error" instead of a "200 OK" and an empty file.

* I decided to introduce a `CloudStorage` class as a namespace for the
base class of the `NotConfiguredError` exception to keep the
`ApplicationController` agnostic of the underlying cloud storage
implementation.

* The "500 Internal server error" status seemed the most appropriate.
I also considered the following:

  * "501 Not implemented", but the functionality *is* implemented, it's
  just not correctly configured.
  * "503 Service unavailable", but this error supposed to only be
  temporary and that doesn't seem quite right in this case.
  * "404 Not found", but the problem seems more serious than that.

* I decided to capitalise "Internal" in the error status, because that
seems like a more common approach. I wanted to capitalise the
"not found" status for the 404 error, but I was concerned there might be
clients depending on it, so I left it unchanged.

* I decided to append the exception message to the error status string,
because I think it's useful to make as much information available as
possible. I tried to see whether there was any standard way of doing
this for other GOV.UK APIs, but couldn't find any consistent approach.

* I introduced the `disable_cloud_storage_stub: true` option for specs
so that I could write an example where the exception was raised.

Fixes #78.