Add env var switch to stream all assets from S3
If the new `STREAM_ALL_ASSETS_FROM_S3` environment variable is present
then all assets should be streamed from S3, i.e. as if the
`stream_from_s3` request parameter was set. The latter still works if
the environment variable is not set, but has no effect if the
environment variable is set.