Fix Bundler/DuplicatedGem rubocop violation
I want to upgrade rubocop to v0.49.1 and this rule was being violated
by the if/else statement around the declaration of the dependency on
the `gds-sso` gem. While it would be legitimate to argue that there is
no actual duplication in this case, I decided it was simpler to remove
the `BUNDLE_DEV` option altogether. If a developer needs it, it would
be easy enough to add back in temporarily in their development
environment.