Fix Bundler/OrderedGems rubocop violation
I want to upgrade rubocop to v0.49.1 and this rule was being violated
by the dependencies in the `Gemfile` not being listed in alphabetic
order.

Some might argue against this approach, but I'm inclined to go with the
rubocop defaults unless there's a good reason not to. And in any case
that deviation from the default should be codified in the `govuk-lint`
gem.