Exclude specs from Metrics/BlockLength rubocop rule
The default restriction of 25 lines per block doesn't play well with the
DSL style of RSpec as noted here [1].

Initially I investigated adding `rubocop-rspec` [2] to the `govuk-lint`
gem, but this then introduced a whole load more violations.

The change in this commit is a stop-gap solution. I plan to open an
issue on the `govuk-lint` repo to highlight the problem.

[1]: https://stackoverflow.com/questions/40934345/rubocop-25-line-block-size-and-rspec-tests#answer-41187163
[2]: https://github.com/backus/rubocop-rspec