Upload assets to S3 with Cache-Control header set
We want to be able to redirect Asset Manager requests to the assets
stored in an S3 bucket. However, we want as much of the
externally-visible behaviour to remain the same and so we need to set
the 'Cache-Control' header to the same value as in
`MediaController#download`, i.e. the value generated by
`ActionController::ConditionalGet#expires_in` [1] using
`ApplicationController#set_expiry`.

Note that the header is set on the S3 object at the point where it is
uploaded. See the documentation for `Aws::S3::Object#upload_file` [2].

[1]: http://api.rubyonrails.org/v4.2.7.1/classes/ActionController/ConditionalGet.html#method-i-expires_in
[2]: http://docs.aws.amazon.com/sdkforruby/api/Aws/S3/Object.html#upload_file-instance_method