Upload assets to S3 with Content-Disposition header set
We want to be able to redirect Asset Manager requests to the assets
stored in an S3 bucket. However, we want as much of the
externally-visible behaviour to remain the same and so we need to set
the 'Content-Disposition' header to the same value as in
`MediaController#download`, i.e. the value generated by
`ActionController::DataStreaming#send_data` [1] or
`ActionController::DataStreaming#send_file` [2].

Note that unfortunately I couldn't find any re-usable code in Rails or
Rack to construct the `Content-Disposition` header value, so I've had to
construct it manually in `ContentDispositionConfiguration#header_for`.

[1]: http://api.rubyonrails.org/v4.2.7.1/classes/ActionController/DataStreaming.html#method-i-send_data
[2]: http://api.rubyonrails.org/v4.2.7.1/classes/ActionController/DataStreaming.html#method-i-send_file