Optionally redirect public asset requests to S3
We're currently experimenting with different ways to serve assets and
adding this optional behaviour makes it possible to have the Asset
Manager *redirect* public requests for assets to S3 rather than serving
them from the (NFS) filesystem or streaming them from S3 through the
app.

The new behaviour is not enabled by default and can be enabled for a
single request using a query string parameter or for all requests in an
environment by setting an environment variable. See the changes in the
README for details.

Eventually we're probably going to want to use a "virtual host" style
URL for the assets on S3, but this will require us to setup a DNS CNAME
for an appropriate sub-domain of gov.uk and setup a bucket with the
bucket name matching the fully-qualified domain name. To make it
possible to try out the new redirect behaviour before this is in place,
I've made the `virtual_host` option for `Aws::S3::Object#public_url`
configurable via an environment variable.

Note that I chose not to store the public URL in the database, because
the call to `Aws::S3::Object#public_url` doesn't appear to connect to
the remote API, i.e. it already has all the information it needs to
construct the URL.