TEMP: Serve assets from S3 if they exist
This is a temporary change that will allow us to test redirecting to and
serving assets from S3 on integration.

This commit causes some tests to fail but we're not worrying about that
as we're not expecting this commit to end up in master.

We've tested the change locally and are happy that it results in a
redirect to assets uploaded to S3. Assets not on S3 continue to be
served directly from the filesystem.