Optionally instruct nginx to proxy assets from S3
This complements the change in the asset-manager nginx config[1].

The app sets the `X-Accel-Redirect` header to instruct nginx to proxy
the request for an asset to S3.

We're generating a presigned URL for the assets on S3 so that we can
avoid making them publicly accessible. The presigned URL is valid for 1
minute which should be enough time for nginx to use the URL to make the
request.

[1]: https://github.com/alphagov/govuk-puppet/pull/6280