Upgrade airbrake gem v4.0.0 -> v4.3.8
I changed the constraint in the `Gemfile` from `~> 4.0.0` to `~> 4.0`
and ran the following command:

    bundle update --conservative airbrake

My main motivation for doing this is so that `airbrake` reports
`ActionController::BadRequest` exceptions correctly [1]. However, I've
chosen to upgrade to the latest 4.x version, because that should include
other bug fixes without breaking compatibility with `errbit`. I'm pretty
sure this is safe to do, because I can see that a bunch of other GOV.UK
apps are using this version of `airbrake` in production, e.g.

* `calculators`
* `hmrc-manuals-api`
* `manuals-publisher`
* `publisher`
* `rummager`

Closes #122.

[1]: airbrake/airbrake#271