Make order of asset response strategies more logical
I've changed the conditions in the `if/elsif/else` statement to be in a
more logical order. I've then changed all other references to the
strategies to match this order, or that order in reverse as appropriate.

The order reflects some sense of progress from the current solution of
using the "X-Accel-Redirect" mechanism to instruct Nginx to serve the
files from the NFS mount, to the preferred solution of redirecting asset
requests to S3.