Proxy percentage of asset requests to S3 via Nginx
The motivation behind this commit is to allow us to incrementally roll-out
the use of the proxying to S3 via Nginx mechanism. Although we've done a
bunch of performance testing and analysis of the production logs, it's
difficult to be certain that the production stack will be able to handle
all its usual traffic when this mechanism is enabled. By only enabling it
gradually we hope to observe any problems before they become catastrophic.

The `PROXY_ALL_ASSET_REQUESTS_TO_S3_VIA_NGINX` environment variable has
been renamed to `PROXY_PERCENTAGE_OF_ASSET_REQUESTS_TO_S3_VIA_NGINX` and
this should be set to an integer value between 0 and 100.

`Random#rand(100)` is used to generate a random integer between 0 and 99
inclusive for every request. If the random number is strictly less than
the percentage specified in the environment variable then the request is
proxied to S3 via Nginx; otherwise Nginx is instructed to serve the asset
from the NFS mount as usual. This should mean that if the percentage is
set to 25, approximately 25% of requests should be proxied to S3 via Nginx.