Send ETag & Last-Modified response headers when proxying to S3 via Nginx
We're planning to gradually rollout the implementation whereby the Rails app
instructs Nginx to proxy asset requests to S3. The current default
implementation is for the Rails app to instruct Nginx to serve the asset files
from NFS. In this implementation Nginx is automatically adding an ETag and a
Last-Modified header.

When we switch to proxying the asset requests to S3, we want to ensure these
response headers are the same for a given asset file in order to avoid
unnecessary cache invalidation and therefore extra load which will make it
hard to discriminate between changes in performance due to the change in
implementation versus changes in performance due to a reduction in caching.

This commit changes the Rails app so that it calculates the ETag and
Last-Modified values based on the underlying file, in the same way that Nginx
already does, and sends them in the response if we are proxying to S3. We plan
to make changes to the Nginx configuration to store these headers and add them
to the proxied response from S3. Without those changes, these response
headers are ignored by Nginx and so I think it's safe to make this change in
isolation.

The ETag value is calculated based on the string format in the
`ngx_http_set_etag` function [1], "%xT-%xO" in combination with the file's
last modified time and its size. It turns out that the "T" and the "O" in
that string format are Nginx-specific and are defined in the `ngx_vslprintf`
function [2]. As far as I can see they are just type modifiers for `time_t` [3]
and `off_t` [4] respectively and I don't think they are relevant to the Ruby
string format, "%x-%x".

According to RFC2616, the ETag value should be quoted [5] and the Last-Modified
value should be formatted as an "HTTP date" [6].

This change means that when we're proxying to S3, the Rails app will have to
access the underlying file on the NFS mount which might slow things down. I've
tried to minimize this by memo-izing the call to `File.stat` which is used to
obtain the data needed to calculate the two header values. This means that the
app should only make a single call to `File.stat` per request.

Note that this solution relies on the fact that the Rails app has access to the
NFS mount. Since the purpose of the change in overall implemenation is to stop
using NFS, we may want to store these header values in the database, but that
can be a separate change.

[1]: http://lxr.nginx.org/source/src/http/ngx_http_core_module.c?v=nginx-1.4.6#1801
[2]: http://lxr.nginx.org/source/src/core/ngx_string.c?v=nginx-1.4.6#0147
[3]: http://lxr.nginx.org/source/src/core/ngx_string.c?v=1.4.6#0279
[4]: http://lxr.nginx.org/source/src/core/ngx_string.c?v=1.4.6#0269
[5]: https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.19
[6]: https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.29