Avoid duplicate Cache-Control headers when proxying to S3 via Nginx
We are setting a Cache-Control response header on the S3 object and
that's the one we want the client to receive. However prior to this
commit, the Rails app was also sending a (duplicate) Cache-Control
response header which Nginx was including in the response. Thus the user
was receiving two duplicate Cache-Control headers. Although the values
would almost certainly have been the same, we want to avoid confusion
and/or unexpected behaviour.

This commit aims to achieve that by stopping the Rails app from sending
a Cache-Control response header. Unfortunately that turns out to be
harder than one might expect and necessitates the following changes:

* Avoid explicitly calling
  `ActionController::ConditionalGet::ClassMethods#expires_in` via
  `ApplicationController#set_expiry`. I think there's an argument for
  not calling this in some of the other cases, but I'm ignoring that for
  now.

* Avoid setting Etag or Last-Modified response headers. Setting either
  of these means that
  `ActionDispatch::Http::Cache#handle_conditional_get!` [1] sets a
  default Cache-Control header value of "max-age=0, private,
  must-revalidate" [2]. We avoid setting ETag or Last-Modified headers
  by renaming them to X-Accel-ETag and X-Accel-Last-Modified. This makes
  sense to me because the response from the Rails app is just an
  instruction to Nginx to proxy the request to S3; Nginx can then make
  use of these headers as it sees fit. Note that this will require a
  corresponding change to the Nginx configuration.

* Remove default Cache-Control response header set by `Rack::ETag`.
  Rails explicitly sets a default Cache-Control header value of
  "no-cache" on this middleware [3] which is added if no ETag header is
  set [4]. By inserting a new custom middleware, we can remove this
  Cache-Control header if the response is instructing Nginx to proxy the
  request to S3. We can detect this scenario via the presence of the
  X-Accel-ETag header in the response.

I considered the following alternative solutions:

1. Change the Nginx configuration to ignore the Cache-Control header
   from Rails.
2. Change the Nginx configuration to ignore the Cache-Control header
   from S3.
3. Remove the Cache-Control header from the S3 object.

While all of these would've avoided the client receiving duplicate
Cache-Control headers, I couldn't find a way to implement option 1 and
options 2 & 3 feel like a step backwards on the journey to redirecting
asset requests to S3. Even if there was a way to implement option 1, it
feels better to me that the Rails app doesn't send the Cache-Control
header in the first place rather than having Nginx ignore it.

[1]:
https://github.com/rails/rails/blob/v4.2.7.1/actionpack/lib/action_dispatch/http/cache.rb#L129-L133
[2]:
https://github.com/rails/rails/blob/v4.2.7.1/actionpack/lib/action_dispatch/http/cache.rb#L135
[3]:
https://github.com/rails/rails/blob/v4.2.7.1/railties/lib/rails/application/default_middleware_stack.rb#L58
[4]: https://github.com/rack/rack/blob/1.6.4/lib/rack/etag.rb#L39