Fail fast if S3 bucket name is not configured
The app will now fail fast with a KeyError is the
`AWS_S3_BUCKET_NAME` is not set in development or production
environments.

In the test environment, `AssetManager.aws_s3_bucket_name` is set to an
easily identifiable value.

I've updated the README to reflect these changes and indicate which
environment variables are required.

Fixes #154.