Store Assets on S3 using Asset#uuid as the object key
Before this commit we were using Asset#id to generate the `key` used
to store the asset in the S3 bucket. We feel that this couples the
storage of the assets a little too closely to Mongo and its mechanism
for generating ids and may make it difficult to change database
technology in the future. In addition it avoids overloading further
the `_id` field in the asset document which is already used to
generate URLs in the API response and filenames for NFS storage.

Instead we'll use a UUID as the object key, generated using Ruby's RFC
4122[1] compliant `SecureRandom#uuid`. This means that the key
generation is in the application code and decoupled from the
database.

At this stage we are not considering changing how assets are stored on
the NFS filesystem, or the object ID created for them in the JSON
response.

There is some additional background on strategies for generating
custom ids on the MongoDB blog[2].

Fixes: #94

[1] https://tools.ietf.org/html/rfc4122
[2] https://www.mongodb.com/blog/post/generating-globally-unique-identifiers-for-use-with-mongodb.