Send more response headers when proxying to S3 via Nginx
And stop setting them on the S3 object.

Prior to this commit, the Rails app was both setting a Cache-Control
response header on the S3 object *and* sending a Cache-Control response
header in the response to Nginx instructing it to proxy the request to
S3. Thus the user was receiving two duplicate `Cache-Control` headers.
Although the values would have been the same, we want to avoid
confusion and/or unexpected behaviour.

After this commit we no longer set the Cache-Control,
Content-Disposition, or Content-Type headers on the S3 object. Instead
the Rails app sends these headers to Nginx. We plan to update the Nginx
configuration in a separate pull request against govuk-puppet in order
to ensure the headers from Rails are added to the proxied response.

In #158 we considered preventing the Rails app from sending the
Cache-Control header, but as @chrisroos pointed out, it was confusing
that some headers were being set by Rails and some by S3. Given that we
want to override the ETag and Last-Modified response headers to match
those sent by Nginx when using the default Sendfile functionality (see
Rails app to be the canonical source of *all* response headers when
proxying to S3 via Nginx.

Note that the Cache-Control header is already being set via the call to
`ApplicationController#set_expiry` and so we don't need to set it
more verbosely like we have for the other headers.

Addresses the Asset Manager aspect of #149. As noted above, we will need
to make some corresponding changes to govuk-puppet.