Add a Rake task to migrate assets to have uuids
The `migrate:add_uuid_to_assets` task added in this commit adds a
`uuid` to all Asset records that do not currently have one. The `uuid`
field was added in the previous commits. Although it is not a required
field, making sure it is pre-populated means it will be easier to use
it in the future. We just want all Assets to have a `uuid` so we use
`update_attribute` to skip any validation.

I ran this task on my development machine against a full clone of the
`govuk_assets_production.assets` table. It updated 58,541 records in
1m22s. I'm happy that this will be representative of the time taken on
production environments.

`Mongoid::Criteria` (returned by a `where` query) enumerates lazily as
standard so we don't need to `find_in_batches` or similar when
iterating over all Assets[1]

I considered adding this as a migration using the
`mongoid_rails_migrations` gem but we've had some problems with that
gem and upgrades in the past[2] and it doesn't seem worth introducing
another dependency for this one-off migration. In development most
developers will use a copy of production data, or be creating new
Assets so the migration won't need to be run.

[1]
https://jira.mongodb.org/browse/MONGOID-1335?focusedCommentId=977653&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-977653
[2] https://github.com/alphagov/manuals-publisher/commit/598f6692a0e805dbc0625a5bf4329a298c728d42