Fail fast if S3 bucket name is not configured in production
The app will now fail fast with a KeyError if the `AWS_S3_BUCKET_NAME`
environment variable is not set in the production environment.

If the environment variable is not set in the development environment
then a "null object" version of the `S3Storage` class is used so that
uploads are a no-op. However, note that asset requests using the
`proxy_to_s3_via_nginx` option will fail with a 500 Server Error via the
`S3Storage::NotConfiguredError` exception.

In the test environment, by default the `Services.cloud_storage` method
returns an RSpec stub. This can be overridden by setting
`disable_cloud_storage_stub: true` in the RSpec metadata for an example,
context, or spec.

I've updated the README to reflect these changes and indicate which
environment variables are required in production.

Fixes #154.