Avoid sending empty Content-Type header when proxying to S3
Prior to this commit, `Asset#content_type` would return an empty string
if the file extension wasn't found by `Mime::Type.lookup`. This meant
that the `Content-Type` response header was also being set to an empty
string. Something else in the Rails stack (presumably some middleware)
was then replacing this empty string with "*/*; charset=utf-8". The
effect of this was that the `Content-Type` response header for files
with unrecognised mime types would be 'application/octet-stream' when
serving the file from the disk and '*/*; charset=utf-8' when serving the
file from S3.

This commit ensures we send the default content type
('application/octet-stream') for files with unrecognised mime types when
proxying requests to S3.

Note that `Mime::Type.lookup_by_extension(nil)` returns the same as
`Mime::Type.lookup_by_extension('unrecognised-extension')` which allows
me to get rid of the `if extension.present?` condition.

Fixes issue #175.