Do not upload asset to S3 if file content is identical
We're planning to upload all existing assets to S3 and we want to avoid
uploading the same asset multiple times if, for example, we have to
repeat the exercise. This will avoid unnecessary use of bandwidth and
will save time.

We can avoid uploading an identical asset to S3 multiple times by storing
an MD5 hex digest as custom metadata on the S3 object and then making a
single HEAD request to detect whether the file already exists on S3 and
whether it is up-to-date.

We considered using the built-in ETag, but it's not straightforward to
compute the ETag for an S3 object, because the rules are different when
the object has been uploaded using multi-part uploads.

We also noticed the built-in Content-MD5 option, but this appears to
only be used to validate the content on upload and is then discarded,
i.e. it cannot be read from S3 after the upload.

Unfortunately there isn't a method on `Aws::S3::Object` to make a HEAD
request. Also the `Aws::S3::Client#head_object` method that we have used
raises an exception if the object does not exist. Both of these things
make the code a bit messier than we would like.

Note that `Aws::S3::Errors::NotFound` requires two arguments for its
constructor (context & message) - the values of these arguments are
irrelevant to what we're testing so we're happy to set them to `nil`.

Note that we are relying on the presence of the file on disk in order to
compute the MD5 hex digest. We will need to re-visit this if/when we
remove the NFS mount.