Add a sidekiq config file
The value for `concurrency` here matches that given in the
`govuk_sidekiq` README[1]. Looking at what other apps in `alphagov`
use[2] it appears that some use this suggested value and others have
tweaked it. 2 seems like a good value to start with, but we may want
to change it.

As far as I can tell, `delayed_job` is currently configured (using
puppet[3]) to run a single worker process on each machine. I'm happy
to allow sidekiq to have some extra concurrency though as this still
feels fairly conservative.

[1] https://github.com/alphagov/govuk_sidekiq
[2] https://github.com/search?p=2&q=org%3Aalphagov+concurrency&type=Code
[3]
https://github.com/alphagov/govuk-puppet/blob/f5d6e85757c1bfd592f08d656cde3527281bd4de/modules/govuk/manifests/apps/asset_manager.pp#L45