Introduce API for adding Whitehall legacy assets
As a step towards having the Asset Manager serve all GOV.UK assets,
we're adding this new API endpoint which allows the Whitehall app to
specify the public URL at which the asset will be made available.

By adding a separate endpoint including the word "whitehall" and by
naming the extra parameter `legacy_url_path`, we've tried to make it
clear that this should not be used by other publishing applications.

Note that the app is not yet able to actually serve these Whitehall
legacy assets. We plan to introduce that functionality in later commits.

Note also that we've intentionally *not* added the
`restrict_request_format` before filter to the new controller. See #197
for details.