Add happy path for serving Whitehall assets
Although in Whitehall's `PublicUploadsController#show` `send_file` is
called with an explicit content type (if one is obtainable from the file
extension), we've elected to rely on the default behaviour of
`send_file` which we think it equivalent.

We've chosen to use the content disposition set in the app's
configuration which is currently set to 'inline', because it matches the
value used in Whitehall's `PublicUploadsController`.

We haven't yet completely replicated all the behaviour of Whitehall's
`PublicUploadsController`, but we plan to do that in subsequent commits.